backend/requirements.txt
asgiref==3.8.1
Django==4.2.18
django-filter==24.3
djangorestframework==3.15.2
